requirements.txt
asgiref==3.2.10
Django==3.1.3
django-extensions==3.0.9
django-hashid-field==3.1.3
django-phonenumber-field==5.0.0
django-widget-tweaks==1.4.8
hashids==1.3.1
phonenumbers==8.12.12
pytz==2020.1
sqlparse==0.3.1
psycopg2==2.8.6
djangorestframework==3.12.2
markdown==3.3.3
django-filter==2.4.0
django-debug-toolbar==2.2
django-simple-history==2.12.0
